add exercise create

+++ internal/domain/exercise_category/mysql_repository.go
@@ -8,7 +8,7 @@ import (
 
 type Repository interface {
 	Create(ctx context.Context, req *models.ExerciseCategoryCreateRequest, userID int) (*db.ExerciseCategory, error)
-	Get(ctx context.Context, id int64) (*db.ExerciseCategory, error)
+	Get(ctx context.Context, id int) (*db.ExerciseCategory, error)
 	GetByUserAndName(ctx context.Context, userID int, name string) (*db.ExerciseCategory, error)
 	GetByUserID(ctx context.Context, userID int) ([]*db.ExerciseCategory, error)
 	//List(ctx context.Context, req *utils.PaginationRequest) (*models.ExerciseCategoryList, error)
